// lending-pool/types/ethers-contracts/misc/AaveProtocolDataProvider.ts
export declare namespace AaveProtocolDataProvider {
      
    export type TokenDataStruct = {symbol: string, tokenAddress: AddressLike}

    export type TokenDataStructOutput = [symbol: string, tokenAddress: string] & {symbol: string, tokenAddress: string }
  
    }